Refactor: moved Vendor and SoftwareLicense models into new assets app.
