+++ registers/models.py
@@ -273,10 +273,6 @@ class Software(models.Model):
     """
     name = models.CharField(max_length=2048, unique=True)
     url = models.CharField(max_length=2000, null=True, blank=True)
-    license = models.ForeignKey(
-        'registers.SoftwareLicense',
-        on_delete=models.PROTECT,
-        null=True)
     os = models.BooleanField(
         default=False,
         verbose_name='OS',
@@ -679,54 +675,6 @@ class Backup(tracking_models.CommonFields):
 
     class Meta:
         ordering = ('system__name',)
-
-
-class Vendor(models.Model):
-    """Represents the vendor for an external product (software or hardware).
-    """
-    name = models.CharField(max_length=256, unique=True)
-    details = models.TextField(blank=True)
-    extra_data = JSONField(default=dict(), null=True, blank=True)
-
-    class Meta:
-        ordering = ('name',)
-
-    def __str__(self):
-        return self.name
-
-
-class SoftwareLicense(tracking_models.CommonFields):
-    """Represents a software licensing arrangement.
-    """
-    name = models.CharField(max_length=256, unique=True)
-    url = models.URLField(max_length=2000, null=True, blank=True)
-    support = models.TextField(
-        blank=True, help_text='Support timeframe or scope')
-    support_url = models.URLField(max_length=2000, null=True, blank=True)
-    oss = models.NullBooleanField(
-        default=None,
-        help_text='Open-source/free software license?')
-    primary_user = models.ForeignKey(
-        tracking_models.DepartmentUser,
-        on_delete=models.PROTECT,
-        null=True,
-        blank=True)
-    vendor = models.ForeignKey(
-        Vendor,
-        on_delete=models.PROTECT,
-        null=True,
-        blank=True)
-    used_licenses = models.PositiveSmallIntegerField(default=0, editable=False)
-    available_licenses = models.PositiveSmallIntegerField(
-        default=0, null=True, blank=True)
-    license_details = models.TextField(
-        blank=True, help_text='Direct license keys or details')
-
-    class Meta:
-        ordering = ('name',)
-
-    def __str__(self):
-        return self.name
 
 
 class BusinessService(models.Model):
